dealing with qemu on the fly dtb mods
So apparently qemu modifies the binary from of the dtb on the fly

- https://github.com/qemu/qemu/blob/v2.11.1/hw/arm/vexpress.c#L468

to get virtio functionality ....

So in the case where we load our own dtb from u-boot (originally was the
vanilla dtb from the kernel sources), we need a dtb that is the result
of this binary on the fly modification process. I injected some snooping
code into QEMU to figure out exactly what this dtb needs to look like
and then used that information to derive the dtb that now has the .qemu
suffix. u-boot can now boot a virtio-capable kernel.