running with dtb extracted directly from qemu
in reading the qemu code, it turns out one can extract the entire dtb by
adding the -machine parameter dumpdtb=<filename>. Did this, disassembled
the result from dtb->dts shoved it in the linux kernel tree and built
the 4.15 kernel dtb for the arm a9 and viola the entire network
subsystem is now working (before just 1 interface was coming up).

I suspect what was not working before was something to do with the
interrupt maps. The dts I made from hand by observing qemu execution
traces is in the *.qemu.src file. I would still like to figure out what
the key discrepency between that file and the one grabbed from the
dumpdtb is.